requirements.txt
attrs==20.3.0
aws-cdk.assets==1.76.0
aws-cdk.aws-apigatewayv2==1.76.0
aws-cdk.aws-apigatewayv2-integrations==1.76.0
aws-cdk.aws-applicationautoscaling==1.76.0
aws-cdk.aws-autoscaling-common==1.76.0
aws-cdk.aws-certificatemanager==1.76.0
aws-cdk.aws-cloudwatch==1.76.0
aws-cdk.aws-codeguruprofiler==1.76.0
aws-cdk.aws-ec2==1.76.0
aws-cdk.aws-ecr==1.76.0
aws-cdk.aws-ecr-assets==1.76.0
aws-cdk.aws-efs==1.76.0
aws-cdk.aws-elasticloadbalancingv2==1.76.0
aws-cdk.aws-events==1.76.0
aws-cdk.aws-iam==1.76.0
aws-cdk.aws-kms==1.76.0
aws-cdk.aws-lambda==1.76.0
aws-cdk.aws-lambda-python==1.76.0
aws-cdk.aws-logs==1.76.0
aws-cdk.aws-route53==1.76.0
aws-cdk.aws-s3==1.76.0
aws-cdk.aws-s3-assets==1.76.0
aws-cdk.aws-sam==1.76.0
aws-cdk.aws-secretsmanager==1.76.0
aws-cdk.aws-servicediscovery==1.76.0
aws-cdk.aws-sqs==1.76.0
aws-cdk.aws-ssm==1.76.0
aws-cdk.cloud-assembly-schema==1.76.0
aws-cdk.core==1.76.0
aws-cdk.cx-api==1.76.0
aws-cdk.region-info==1.76.0
boto3==1.16.28
botocore==1.19.28
cattrs==1.1.2
certifi==2020.11.8
chardet==3.0.4
constructs==3.2.56
idna==2.10
jmespath==0.10.0
jsii==1.15.0
publication==0.0.3
python-dateutil==2.8.1
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
